Better tool display when G92 present
Previously, in programs where G92 is used, e.g. to change
the origin for step-and-repeat milling operations,
the tool image would follow only the first occurrence
of the repeated path.  Now it properly moves to the new
location.  To accomplish this, the tool displayer now
operates in machine coordinates with an offset determined
by the starting point of the program.