Highlight line being executed in gcode viewer
This works well only with g2core, and requires changes to the server:
a) Do not skip the sending of blank lines
b) Add N0 to lines sent by the feeder

If there are no blank lines in the GCode program, (a) is unnecessary.
If (b) is not done, then jogging will cause the GCode display
to scroll to the last line that was executed.